% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{afsc_grid}
\alias{afsc_grid}
\title{Data frame containing multiple survey grids for surveys collected by NOAA's Alaska
Fisheries Science Center. Columns include the latitude
and longitude in decimal degrees, area of each survey cell (in km2), the survey
name (\code{survey}) year of update (\code{survey_domain_year}) and depth (currently NA
for this dataset and must be joined in separately). Survey names generally match
those for data in \code{get_data()}}
\format{
A data frame.
}
\usage{
afsc_grid
}
\description{
Data frame containing multiple survey grids for surveys collected by NOAA's Alaska
Fisheries Science Center. Columns include the latitude
and longitude in decimal degrees, area of each survey cell (in km2), the survey
name (\code{survey}) year of update (\code{survey_domain_year}) and depth (currently NA
for this dataset and must be joined in separately). Survey names generally match
those for data in \code{get_data()}
}
\keyword{datasets}
