% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_version}
\alias{data_version}
\title{Public facing function to create table of metadata}
\usage{
data_version()
}
\value{
dataframe containing file names and dates of last update
}
\description{
Public facing function to create table of metadata
}
