% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfo_synoptic_grid}
\alias{dfo_synoptic_grid}
\title{Data frame containing the survey grid for the Fisheries and Oceans Canada's
synoptic trawl survey collected by DFO's Pacific Biological Station.
Columns include the latitude and longitude in decimal degrees, area of each
survey cell (in hectares), the survey name (\code{survey}) and year of update
(\code{survey_domain_year}) and depth}
\format{
A data frame.
}
\usage{
dfo_synoptic_grid
}
\description{
Data frame containing the survey grid for the Fisheries and Oceans Canada's
synoptic trawl survey collected by DFO's Pacific Biological Station.
Columns include the latitude and longitude in decimal degrees, area of each
survey cell (in hectares), the survey name (\code{survey}) and year of update
(\code{survey_domain_year}) and depth
}
\keyword{datasets}
