% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download}
\alias{download}
\title{Perform downloading}
\usage{
download(x, cache_folder = get_cache_folder())
}
\arguments{
\item{x}{the directory name, default is the package name "surveyjoin"}

\item{cache_folder}{the name of the folder used for caching}
}
\description{
Perform downloading
}
