% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_last_modified}
\alias{file_last_modified}
\title{Function to get the last modified date of a file from GitHub}
\usage{
file_last_modified(file_name)
}
\arguments{
\item{file_name}{the file name, e.g. "nwfsc-catch.rds"}
}
\value{
The time stamp file was last changed
}
\description{
Function to get the last modified date of a file from GitHub
}
