% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cache_folder}
\alias{get_cache_folder}
\title{Identify the local folder used for caching}
\usage{
get_cache_folder(name = "surveyjoin")
}
\arguments{
\item{name}{The directory name, default is the package name \code{"surveyjoin"}.}
}
\value{
The directory location used for caching
}
\description{
Identify the local folder used for caching
}
