% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data}
\alias{get_data}
\title{Query the survey database}
\usage{
get_data(
  common = NULL,
  scientific = NULL,
  itis_id = NULL,
  regions = NULL,
  surveys = NULL,
  years = NULL
)
}
\arguments{
\item{common}{A string, or vector of strings of common names for species}

\item{scientific}{A string, or vector of strings of scientific names for species}

\item{itis_id}{An integer or vector of integers corresponding to ITIS identifiers}

\item{regions}{A string, or vector of strings of common names for regions. May be one or more
of "afsc", "nwfsc", "pbs". Surveys are nested within region, so returning data from a region will
return more than one survey.}

\item{surveys}{A string, or vector of strings of common names for surveys. May be one or more
of "Aleutian Islands Bottom Trawl Survey", "Eastern Bering Sea Crab/Groundfish Bottom Trawl Survey",
"Eastern Bering Sea Slope Bottom Trawl Survey", "Gulf of Alaska Bottom Trawl Survey",
"Northern Bering Sea Crab/Groundfish Survey - Eastern Bering Sea Shelf Survey Extension",
"NWFSC.Combo", "NWFSC.Shelf", "NWFSC.Hypoxia", "NWFSC.Hypoxia", "Triennial", "SYN QCS",
"SYN HS", "SYN WCVI", "SYN WCHG". If NULL, all are returned}

\item{years}{a vector of years, e.g. \code{year = 2013:2018}. If NULL, all are returned}
}
\value{
a dataframe of joined haul and catch data
}
\description{
Query the survey database
}
\examples{
\dontrun{}
}
