% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_itis_spp}
\alias{get_itis_spp}
\title{Get the ITIS identifier of a species}
\usage{
get_itis_spp(spp)
}
\arguments{
\item{spp}{(character) Taxonomic name to query}
}
\value{
an integer representing the ITIS identifier
}
\description{
Get the ITIS identifier of a species
}
\examples{
\dontrun{
id <- get_itis_spp("darkblotched rockfish")

}
}
