% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_rawdata}
\alias{get_rawdata}
\title{Get the URLs to raw data for each dataset}
\usage{
get_rawdata()
}
\value{
a dataframe with the URL for each region
}
\description{
Get the URLs to raw data for each dataset
}
\examples{
\dontrun{
m <- get_rawdata()
}
}
