% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_shapefiles}
\alias{get_shapefiles}
\title{Get the survey grid shapefiles for each dataset}
\usage{
get_shapefiles()
}
\value{
a dataframe with the URL for each region
}
\description{
Get the survey grid shapefiles for each dataset
}
\examples{
\dontrun{
m <- get_shapefiles()
}
}
