% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_species}
\alias{get_species}
\title{Get the table of common and scientific names in the joined dataset}
\usage{
get_species()
}
\value{
a dataframe with the common and scientific name
}
\description{
Get the table of common and scientific names in the joined dataset
}
\examples{
\dontrun{
get_species()
}
}
