% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_survey_names}
\alias{get_survey_names}
\title{Get the name for each survey and corresponding region}
\usage{
get_survey_names()
}
\value{
a dataframe with the survey name and associated region
}
\description{
Get the name for each survey and corresponding region
}
\examples{
\dontrun{
m <- get_survey_names()
}
}
