% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_sql_data}
\alias{load_sql_data}
\title{Load SQLite database}
\usage{
load_sql_data()
}
\value{
Nothing; data is inserted into a local SQLite database.
}
\description{
Load SQLite database
}
\examples{
\dontrun{
load_sql_data()
}
}
