% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nwfsc_catch}
\alias{nwfsc_catch}
\title{Dataframe of all catch data from NWFSC trawl surveys}
\format{
A data frame.
}
\usage{
nwfsc_catch
}
\description{
Dataframe of all catch data from NWFSC trawl surveys
}
\keyword{datasets}
