% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nwfsc_haul}
\alias{nwfsc_haul}
\title{Dataframe of all haul attributes from NWFSC trawl surveys}
\format{
A data frame.
}
\usage{
nwfsc_haul
}
\description{
Dataframe of all haul attributes from NWFSC trawl surveys
}
\keyword{datasets}
