% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_metadata}
\alias{save_metadata}
\title{Function to save metadata}
\usage{
save_metadata(metadata)
}
\arguments{
\item{metadata}{the name of the metadata object}
}
\description{
Function to save metadata
}
