% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species}
\alias{species}
\title{Dataframe of common and scientific names for surveys}
\format{
A data frame.
}
\usage{
species
}
\description{
Dataframe of common and scientific names for surveys
}
\keyword{datasets}
