% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spp_dictionary}
\alias{spp_dictionary}
\title{Data frame of species common names, scientific names, and ITIS numbers}
\format{
A data frame.
}
\usage{
spp_dictionary
}
\description{
Data frame of species common names, scientific names, and ITIS numbers
}
\keyword{datasets}
