% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{surv_db}
\alias{surv_db}
\title{Load the survey database}
\usage{
surv_db()
}
\value{
a \code{\link[DBI:dbConnect]{DBI::dbConnect()}} connection to the database
}
\description{
Load the survey database
}
\examples{
\dontrun{
db <- surv_db()

}
}
