% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cache_data}
\alias{cache_data}
\title{Function to cache the data files locally}
\usage{
cache_data(region = c("nwfsc", "pbs", "afsc"))
}
\arguments{
\item{region}{Region name(s)}
}
\value{
Nothing returned; data files are cached locally.
}
\description{
Function to cache the data files locally
}
\examples{
\dontrun{
cache_data()
}
}
