% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Main function to query the survey database}
\usage{
get_data(
  common = NULL,
  scientific = NULL,
  itis_id = NULL,
  regions = NULL,
  surveys = NULL,
  years = NULL
)
}
\arguments{
\item{common}{A string, or vector of strings of common names for species. This is NULL by default, so that all species are returned.}

\item{scientific}{A string, or vector of strings of scientific names for species. This is NULL by default, so that all species are returned.}

\item{itis_id}{An integer or vector of integers corresponding to ITIS identifiers. This is NULL by default, so that all species are returned.}

\item{regions}{A string, or vector of strings of common names for regions. May be one or more
of "afsc", "nwfsc", "pbs". Surveys are nested within region, so returning data from a region will
return more than one survey. This is NULL by default, so that all surveys are returned.}

\item{surveys}{A string, or vector of strings of common names for surveys. May be one or more
of "Aleutian Islands Bottom Trawl Survey", "Eastern Bering Sea Crab/Groundfish Bottom Trawl Survey",
"Eastern Bering Sea Slope Bottom Trawl Survey", "Gulf of Alaska Bottom Trawl Survey",
"Northern Bering Sea Crab/Groundfish Survey - Eastern Bering Sea Shelf Survey Extension",
"NWFSC.Combo", "NWFSC.Shelf", "NWFSC.Hypoxia", "Triennial", "SYN QCS",
"SYN HS", "SYN WCVI", "SYN WCHG". If NULL, all are returned.}

\item{years}{a vector of years, e.g. \code{year = 2013:2018}. If NULL, all are returned.}
}
\value{
a dataframe of joined haul and catch data
\itemize{
\item \strong{event_id}: Unique haul identifier.
\item \strong{itis}: ITIS identifier for species.
\item \strong{catch_numbers}: Numbers of fish for this haul - species.
\item \strong{catch_weight}: Weight (kg) of fish for this haul - species.
\item \strong{region}: Region this survey originated in ("pbs", "nwfsc", "afsc").
\item \strong{scientific_name}: Scientific name for this species.
\item \strong{common_name}: Common name for this species.
\item \strong{survey_name}: Name of the survey this haul is part of.
\item \strong{date}: String representation of the date, format YYYY-MM-DD.
\item \strong{pass}: Optional pass identifier (1 or 2), only used for NWFSC surveys.
\item \strong{vessel}: Optional unique vessel identifier, not included in all surveys.
\item \strong{lat_start}: Starting latitude of haul, decimal degrees.
\item \strong{lon_start}: Starting longitude of haul, decimal degrees.
\item \strong{lat_end}: Ending latitude of haul, decimal degrees.
\item \strong{lon_end}: Ending longitude of haul, decimal degrees.
\item \strong{depth_m}: Haul bottom depth (meters).
\item \strong{effort}: Amount of units corresponding to this haul.
\item \strong{effort_units}: Units of effort.
\item \strong{performance}: Optional performance indicator for each haul, not used for all surveys. If not
indicated, assume performance is satisfactory
\item \strong{bottom_temp_c}: Bottom temperature recorded at the gear, in degrees Celsius.
\item \strong{year}: Calendar year corresponding to the haul.
}
}
\description{
Main function to query the survey database
}
\examples{
\dontrun{
d <- get_data()
}
}
