% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_metadata}
\alias{get_metadata}
\title{Get the metadata URLs for each dataset}
\usage{
get_metadata()
}
\value{
a dataframe with the metadata URL for each region
}
\description{
Get the metadata URLs for each dataset
}
\examples{
\dontrun{
m <- get_metadata()
}
}
